Fix handling of empty application_name
It was reported in #993 that using an empty `application_name` on
connection startup could result in the actual application_name that was
set later not being detected. The reason was that we were explicitely
not forwarding empty strings from clients as settings to Postgres. This
is fixed by simply removing this check. Why we were explicitely ignoring
empty strings is not clear to me. But our handling of startup parameters
and GUCs has heavily changed over the years, and is able to handle them
fine like this.